+++ src/Core/Types/Configuration/TypeRegistrar.cs
@@ -208,13 +208,21 @@ namespace HotChocolate.Configuration
                 }
                 else
                 {
-                    var normalizedTypeRef = new ClrTypeReference(
-                        typeInfo.ClrType,
-                        typeReference.Context);
-
-                    if (!IsTypeResolved(normalizedTypeRef))
+                    for (int i = 0; i < typeInfo.Components.Count; i++)
                     {
-                        Unresolved.Add(normalizedTypeRef);
+                        var normalizedTypeRef = new ClrTypeReference(
+                            typeInfo.Components[i],
+                            typeReference.Context);
+
+                        if (IsTypeResolved(normalizedTypeRef))
+                        {
+                            break;
+                        }
+
+                        if ((i + 1) == typeInfo.Components.Count)
+                        {
+                            Unresolved.Add(normalizedTypeRef);
+                        }
                     }
                 }
             }
@@ -325,7 +333,6 @@ namespace HotChocolate.Configuration
             }
             return false;
         }
-
 
         private TypeSystemObjectBase CreateInstance(Type type) =>
             (TypeSystemObjectBase)_serviceFactory.CreateInstance(type);