+++ src/Core/Types/Utilities/TypeInfo.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using System.Collections.Generic;
+using System;
 using HotChocolate.Types;
 
 namespace HotChocolate.Utilities
@@ -6,11 +7,15 @@ namespace HotChocolate.Utilities
     internal sealed class TypeInfo
     {
         public TypeInfo(Type clrType,
+            IReadOnlyList<Type> components,
             Func<IType, IType> typeFactory)
         {
             ClrType = clrType;
+            Components = components;
             TypeFactory = typeFactory;
         }
+
+        public IReadOnlyList<Type> Components { get; }
 
         public Type ClrType { get; }